+++ lib/providers/devices_providers.dart
@@ -1,5 +1,3 @@
-import 'dart:async';
-
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
 import 'package:attentio_desktop/src/rust/api/device_api.dart';